Created new timeseries module for view factor calculations
